Make rtimer 32-bit by default

+++ os/sys/rtimer.h
@@ -56,8 +56,9 @@
 #include "contiki.h"
 
 #ifndef RTIMER_CLOCK_DIFF
-typedef unsigned short rtimer_clock_t;
-#define RTIMER_CLOCK_DIFF(a,b)     ((signed short)((a)-(b)))
+/* By default: 32-bit rtimer. Set RTIMER_CLOCK_DIFF for custom rtimer width */
+typedef uint32_t rtimer_clock_t;
+#define RTIMER_CLOCK_DIFF(a, b)     ((int32_t)((a) - (b)))
 #endif /* RTIMER_CLOCK_DIFF */
 
 #define RTIMER_CLOCK_LT(a, b)      (RTIMER_CLOCK_DIFF((a),(b)) < 0)
